+++ src/commands/BuildCommand.js
@@ -20,7 +20,7 @@ import SnippetBuilder from './runners/SnippetBuilder.js'
 class BuildCommand {
   /**
    * Execute Build Command
-   * @returns {Promise<FSWatcher|void>}
+   * @returns {Promise<FSWatcher|null>}
    */
   static async execute () {
     const collectionName = Session.config.name
@@ -31,35 +31,32 @@ class BuildCommand {
       componentNames = [Session.targets]
     }
 
+    const collection = await CollectionFactory.fromName(collectionName, componentNames)
+    await this.buildCollection(collection)
+    await this.deployCollection(collection)
+
     if (Session.watchMode) {
-      const collection = await this.buildCollection(collectionName, componentNames)
-      await this.deployCollection(collection)
       Session.firstRun = false
       return this.watchCollection(collection)
     }
 
-    const collection = await this.buildCollection(collectionName, componentNames)
-    return this.deployCollection(collection)
+    return null
   }
 
   /**
    * Build a Collection
-   * @param {string} collectionName Collection Name
-   * @param {string[]} componentNames Collection Component Names
+   * @param {module:models/Collection} collection Collection
    * @throws InternalError - No components found
    * @return {Promise<module:models/Collection>}
    */
-  static async buildCollection (collectionName, componentNames) {
+  static async buildCollection (collection) {
     // If this is a Theme, the Current Target Name will always be the Collection Name.
     // Let's use that instead of Session.targets which might contain a Collection List object.
     /** @type {string} **/
-    const currentTargetName = Session.callerType === Components.THEME_TYPE_NAME ? collectionName : Session.targets
+    const currentTargetName = Session.callerType === Components.THEME_TYPE_NAME ? collection.name : Session.targets
 
     logTitleItem(`Initializing Components for "${currentTargetName}"`)
     const initStartTime = new Timer()
-
-    // Init Collection
-    let collection = await CollectionFactory.fromName(collectionName, componentNames)
 
     // Initialize Individual Components
     collection.components = await Promise.all(collection.components.map(component => ComponentFactory.initializeComponent(component)))
@@ -75,11 +72,11 @@ class BuildCommand {
     logChildItem(`Found ${collection.components.length} component${plural(collection.components)} and  ${collection.snippets.length} snippet${plural(collection.snippets)}.`)
 
     // Filter Out Components When Applicable
-    if (componentNames?.length) {
+    if (collection.componentNames?.length) {
       // for each component, get tree item names
-      const componentNamesToBuild = CollectionUtils.getComponentsNameTree(allComponents, componentNames)
+      const componentNamesToBuild = CollectionUtils.getComponentsNameTree(allComponents, collection.componentNames)
 
-      logChildItem(`Packaging the following component${plural(componentNames)}: ${componentNames.join(', ')}`)
+      logChildItem(`Packaging the following component${plural(collection.componentNames)}: ${collection.componentNames.join(', ')}`)
 
       collection.components = collection.components.filter(component => componentNamesToBuild.has(component.name))
       collection.snippets = collection.snippets.filter(snippet => componentNamesToBuild.has(snippet.name))
@@ -87,7 +84,7 @@ class BuildCommand {
 
     // Throw an Error when No Components are found
     if (collection.components.length + collection.snippets.length === 0) {
-      throw new InternalError(`No matching components found for [${componentNames.join(',')}]`)
+      throw new InternalError(`No matching components found for [${collection.componentNames.join(',')}]`)
     }
 
     logChildItem(`Initialization complete (${initStartTime.now()} seconds)`)
@@ -117,7 +114,7 @@ class BuildCommand {
     await this.setComponentHierarchy(collection.snippets, allComponents)
 
     logChildMessage()
-    logChildMessage(`${collectionName}/`)
+    logChildMessage(`${collection.name}/`)
 
     let filteredComponents = collection.components.filter(component => component.name.startsWith('section'))
     if (!filteredComponents.length > 0) {
@@ -254,7 +251,8 @@ class BuildCommand {
     const collectionName = Session.config.name
     const componentNames = Session.config?.components
 
-    const collection = await this.buildCollection(collectionName, componentNames)
+    const collection = await CollectionFactory.fromName(collectionName, componentNames)
+    await this.buildCollection(collection)
     await this.deployCollection(collection)
     // Restart Watcher on liquid file change to make sure we do refresh watcher snippet folders
     if (filename.endsWith('.liquid')) {