+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2024-01-31
+
+### Added
+
+- Install functionality is now paired with shopify theme dev functionality.
+
+### Fixed
+
+- Install functionality fully restored
+
 ## [2.1.0] - 2024-01-30
 
 ### Added
@@ -22,25 +32,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add omitted watch flag for install oclif command
 - Add missing trace/debug flags
 
-## [2.0.3]
+## [2.0.3] - 2024-01-29
 
 ### Changed
 
 - Moved @types/node dependency for oclif compliance as a dependency
 
-## [2.0.2]
+## [2.0.2] - 2024-01-29
 
 ### Added
 
 - Added @types/node dependency for oclif compliance as a dev dependency
 
-## [2.0.1]
+## [2.0.1] - 2024-01-29
 
 ### Changed
 
 - License set to MIT to match Shopify CLI license
 
-## [2.0.0]
+## [2.0.0] - 2024-01-29
 
 ### Changed